Move shared export and rpc types into surrealdb-client-core
