+++ surrealdb/client-core/src/lib.rs
@@ -1,1 +1,3 @@
+pub mod export;
+pub mod rpc;
 pub mod response;
